fix: JWT 최적화
- MemberEntity의 복잡도 감소를 위해 JWTEntity 추가로 생성
- JWT는 DB에 접근하지 않아도 사용자의 정보를 가지고 있다는 장점을 이용해 기존 DB에 접근하던 방식에서 벗어남
