+++ src/main/java/keeper/project/homepage/common/service/sign/SignInService.java
@@ -103,9 +103,10 @@ public class SignInService {
   private boolean passwordMatches(String password, String hashedPassword) {
     boolean isPasswordMatch = false;
     try {
-    isPasswordMatch = customPasswordService.checkPasswordWithPBKDF2SHA256(password, hashedPassword)
-        || customPasswordService.checkPasswordWithMD5(password, hashedPassword)
-        || passwordEncoder.matches(password, hashedPassword);
+      isPasswordMatch =
+          customPasswordService.checkPasswordWithPBKDF2SHA256(password, hashedPassword)
+              || customPasswordService.checkPasswordWithMD5(password, hashedPassword)
+              || passwordEncoder.matches(password, hashedPassword);
     } catch (Exception e) {
       throw new CustomLoginIdSigninFailedException("비밀번호가 일치하지 않습니다.");
     }
