+++ src/main/java/keeper/project/homepage/config/security/JwtMemberEntity.java
@@ -1,0 +1,61 @@
+package keeper.project.homepage.config.security;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.security.Principal;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.security.auth.Subject;
+import keeper.project.homepage.entity.member.MemberHasMemberJobEntity;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+
+@RequiredArgsConstructor
+public class JwtMemberEntity implements UserDetails {
+
+  private final String id;
+  private final List<String> memberJobs;
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    List<SimpleGrantedAuthority> roles = new ArrayList<>();
+
+    for (String memberJob : memberJobs) {
+      roles.add(new SimpleGrantedAuthority(memberJob));
+    }
+    return roles;
+  }
+
+  @Override
+  public String getPassword() {
+    return "";
+  }
+
+  @Override
+  public String getUsername() {
+    return id;
+  }
+
+  @Override
+  public boolean isAccountNonExpired() {
+    return true;
+  }
+
+  @Override
+  public boolean isAccountNonLocked() {
+    return true;
+  }
+
+  @Override
+  public boolean isCredentialsNonExpired() {
+    return true;
+  }
+
+  @Override
+  public boolean isEnabled() {
+    return true;
+  }
+}
