+++ src/main/java/keeper/project/homepage/entity/member/MemberEntity.java
@@ -42,7 +42,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 @NoArgsConstructor          // 인자없는 생성자를 자동으로 생성합니다.
 @AllArgsConstructor         // 인자를 모두 갖춘 생성자를 자동으로 생성합니다.
 @Table(name = "member")     // 'member' 테이블과 매핑됨을 명시
-public class MemberEntity implements UserDetails, Serializable {
+public class MemberEntity implements Serializable {
 
   @Id // pk
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -160,51 +160,6 @@ public class MemberEntity implements UserDetails, Serializable {
   @Builder.Default
   private List<PostingEntity> posting = new ArrayList<>();
 
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    List<SimpleGrantedAuthority> roles = new ArrayList<>();
-
-    for (MemberHasMemberJobEntity memberJob : this.getMemberJobs()) {
-      roles.add(new SimpleGrantedAuthority(memberJob.getMemberJobEntity().getName()));
-    }
-    return roles;
-  }
-
-  @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-  @Override
-  public String getPassword() {
-    return password;
-  }
-
-  @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-  @Override
-  public String getUsername() {
-    return String.valueOf(this.id);
-  }
-
-  @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-  @Override
-  public boolean isAccountNonExpired() {
-    return true;
-  }
-
-  @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-  @Override
-  public boolean isAccountNonLocked() {
-    return true;
-  }
-
-  @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return true;
-  }
-
-  @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-  @Override
-  public boolean isEnabled() {
-    return true;
-  }
 
   public void updatePoint(int point) {
     this.point = point;
